refactor Blocks and DefaultTheme twig extensions to runtime implementations
